Внёс поправки для корректного изменения характеристик товара в листе закупки

+++ CurrentProduct.cs
@@ -54,6 +54,7 @@ namespace Trading_Company
                 bDeleteProduct.Visible = false;
                 bChange.Visible = false;
             }
+
             this.id = id;
             this.id_purchase = id_purchase;
             this.count = count;
@@ -77,7 +78,7 @@ namespace Trading_Company
             unitBox.Text = param.ExecuteScalar().ToString();
             unit = unitBox.Text;
 
-            summary = richSummaryBox.Text;
+            
 
             countBox.Text = count.ToString();
 
@@ -86,12 +87,15 @@ namespace Trading_Company
             param = new SQLiteCommand("SELECT image FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
 
             byte[] image = (byte[])param.ExecuteScalar();
-            pictureBox1.Image = byteArrayToImage(image);
+            
             photo = image;
+            pictureBox1.Image = byteArrayToImage(photo);
             basePhoto = photo;
 
             param = new SQLiteCommand("SELECT characteristic FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
             richSummaryBox.Text = param.ExecuteScalar().ToString();
+
+            summary = richSummaryBox.Text;
 
             ConnectionToDB.closeDB();
         }
@@ -117,6 +121,7 @@ namespace Trading_Company
             {
                 var filePath = string.Empty;
                 OpenFileDialog openFileDialog = new OpenFileDialog();
+                openFileDialog.Filter = "Изображение. Файл формата  *.png|*.jpg|*.jpeg|*.bmp";
                 if (openFileDialog.ShowDialog() == DialogResult.OK)
                 {
                     //Get the path of specified file
@@ -143,33 +148,68 @@ namespace Trading_Company
 
         private void bOk_Click(object sender, EventArgs e)
         {
-            if (nameBox.Text != "" && categoryBox.Text != "" && unitBox.Text != "" && vatBox.Text != "" && countBox.Text != "" && priceBox.Text != "")
+            Boolean close = false;
+
+            if (nameBox.Text != "" && categoryBox.Text != "" && unitBox.Text != "" && vatBox.Text != "" && countBox.Text != "" && priceBox.Text != "" && richSummaryBox.Text != "")
             {
-                if (!(nameBox.Text.Equals(name) && categoryBox.Text.Equals(category) && unitBox.Text.Equals(unit) && vatBox.Text.Equals(vat) && countBox.Text.Equals(count.ToString()) && priceBox.Text.Equals(price.ToString()) && richSummaryBox.Text.Equals(summary) && Enumerable.SequenceEqual(photo, basePhoto)))
+                if ((nameBox.Text.ToLower().Equals(name.ToLower()) && categoryBox.Text.ToLower().Equals(category.ToLower()) && unitBox.Text.ToLower().Equals(unit.ToLower()) && vatBox.Text.Equals(vat) && countBox.Text.Equals(count.ToString()) && priceBox.Text.Equals(price.ToString()) && richSummaryBox.Text.ToLower().Equals(summary.ToLower()) && Enumerable.SequenceEqual(photo, basePhoto)))
+                {
+                    MessageBox.Show("Изменений не нашлось!", "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+                else
                 {
                     ConnectionToDB.openDB();
 
+                    if ((count != Convert.ToInt32(countBox.Text) || price != Convert.ToInt32(priceBox.Text)) && (nameBox.Text.ToLower().Equals(name.ToLower()) && categoryBox.Text.ToLower().Equals(category.ToLower()) && unitBox.Text.ToLower().Equals(unit.ToLower()) && vatBox.Text.Equals(vat) && richSummaryBox.Text.ToLower().Equals(summary.ToLower()) && Enumerable.SequenceEqual(photo, basePhoto)))
+                    {
+                        SQLiteCommand update = new SQLiteCommand("UPDATE product_list_in_purchase SET count_product_in_purchase  = '"+ countBox.Text + "', price_product_in_purchase  = '" + priceBox.Text + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
+                        update.ExecuteNonQuery();
+
+                        ConnectionToDB.closeDB();
+                        close = true;
+                    }
+
+                    SQLiteCommand isnew = new SQLiteCommand("SELECT isnewproduct FROM product_list_in_purchase WHERE id_product = '" + id + "'", ConnectionToDB.DB);
+                    if(!close && isnew.ExecuteScalar().ToString() == "False")
+                    {
+                        SQLiteCommand command = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image, id_company) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
+                    "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo, (SELECT id_company FROM purchase WHERE id = '" + id_purchase + "'))", ConnectionToDB.DB);
+                        command.Parameters.AddWithValue("@photo", photo);
+                        command.ExecuteNonQuery();
+
+                        SQLiteCommand getID = new SQLiteCommand("SELECT MAX(id) FROM products", ConnectionToDB.DB);
+                        int id_prod = Convert.ToInt32(getID.ExecuteScalar());
+
+                        SQLiteCommand create = new SQLiteCommand("INSERT INTO warehouse (id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
+                        create.ExecuteNonQuery();
+
+                        SQLiteCommand update = new SQLiteCommand("UPDATE product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
+                        update.ExecuteNonQuery();
+
+                        ConnectionToDB.closeDB();
+
+                    }
 
                     // А если пользователь ничего не изменил?!?!?!?!
 
                     SQLiteCommand ifCMD = new SQLiteCommand("SELECT count_product FROM warehouse WHERE id_product = '" + id + "'", ConnectionToDB.DB);
 
-                    if (Convert.ToInt32(ifCMD.ExecuteScalar()) == count)
+                    if (!close && Convert.ToInt32(ifCMD.ExecuteScalar()) == count)
                     {
                         SQLiteCommand delete = new SQLiteCommand("DELETE FROM warehouse WHERE id_product = '" + id + "'", ConnectionToDB.DB);
                         delete.ExecuteNonQuery();
                         delete = new SQLiteCommand("DELETE FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
                         delete.ExecuteNonQuery();
 
-                        SQLiteCommand create = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image, id_company) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
-                        "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo, (SELECT id_company FROM purchase WHERE id = '"+ id_purchase +"'))", ConnectionToDB.DB);
-                        create.Parameters.AddWithValue("@photo", photo);
-                        create.ExecuteNonQuery();
+                        SQLiteCommand command = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image, id_company) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
+                    "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo, (SELECT id_company FROM purchase WHERE id = '" + id_purchase + "'))", ConnectionToDB.DB);
+                        command.Parameters.AddWithValue("@photo", photo);
+                        command.ExecuteNonQuery();
 
                         SQLiteCommand getID = new SQLiteCommand("SELECT MAX(id) FROM products", ConnectionToDB.DB);
                         int id_prod = Convert.ToInt32(getID.ExecuteScalar());
 
-                        create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
+                        SQLiteCommand create = new SQLiteCommand("INSERT INTO warehouse (id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
                         create.ExecuteNonQuery();
 
                         SQLiteCommand update = new SQLiteCommand("UPDATE product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
@@ -177,34 +217,36 @@ namespace Trading_Company
                     }
                     else
                     {
-                        SQLiteCommand delete = new SQLiteCommand("UPDATE FROM warehouse SET count_product = count_product - '" + count + "'", ConnectionToDB.DB);
-                        delete.ExecuteNonQuery();
+                        if (!close)
+                        {
+                            SQLiteCommand delete = new SQLiteCommand("UPDATE warehouse SET count_product = count_product - '" + count + "'", ConnectionToDB.DB);
+                            delete.ExecuteNonQuery();
 
-                        SQLiteCommand create = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image, id_company) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
-                            "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo, (SELECT id_company FROM purchase WHERE id = '" + id_purchase + "')", ConnectionToDB.DB);
-                        create.Parameters.AddWithValue("@photo", photo);
-                        create.ExecuteNonQuery();
+                            SQLiteCommand command = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image, id_company) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
+                        "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo, (SELECT id_company FROM purchase WHERE id = '" + id_purchase + "'))", ConnectionToDB.DB);
+                            command.Parameters.AddWithValue("@photo", photo);
+                            command.ExecuteNonQuery();
 
-                        SQLiteCommand getID = new SQLiteCommand("SELECT id(MAX) FROM products", ConnectionToDB.DB);
-                        int id_prod = Convert.ToInt32(getID.ExecuteScalar());
+                            SQLiteCommand getID = new SQLiteCommand("SELECT MAX(id) FROM products", ConnectionToDB.DB);
+                            int id_prod = Convert.ToInt32(getID.ExecuteScalar());
 
-                        create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
-                        create.ExecuteNonQuery();
+                            SQLiteCommand create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
+                            create.ExecuteNonQuery();
 
-                        SQLiteCommand update = new SQLiteCommand("UPDATE FROM product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
-                        update.ExecuteNonQuery();
+                            SQLiteCommand update = new SQLiteCommand("UPDATE product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
+                            update.ExecuteNonQuery();
+                        }
                     }
 
                     ConnectionToDB.closeDB();
                 }
-                else
-                {
-                    MessageBox.Show("Изменений не нашлось!", "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
             }
+            
             else
             {
                 MessageBox.Show("Все поля должны быть заполнены", "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error);
+
+
             }
             this.Close();
         }
@@ -218,8 +260,25 @@ namespace Trading_Company
         {
             ConnectionToDB.openDB();
 
-            SQLiteCommand delProd = new SQLiteCommand("DELETE FROM product_list_in_purchase WHERE id_product = '" + id + "' AND id_purchase = '"+ id_purchase +"' AND price_product_in_purchase = '"+ price +"' AND count_product_in_purchase = '"+ count +"'", ConnectionToDB.DB);
-            delProd.ExecuteNonQuery();
+            SQLiteCommand ifCMD = new SQLiteCommand("SELECT isnewproduct FROM product_list_in_purchase WHERE id_product = '"+ id + "' AND id_purchase = '" + id_purchase + "' AND price_product_in_purchase = '" + price + "' AND count_product_in_purchase = '" + count + "'", ConnectionToDB.DB);
+
+            if(ifCMD.ExecuteScalar().ToString() == "True")
+            {
+                SQLiteCommand delProd = new SQLiteCommand("DELETE FROM product_list_in_purchase WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND price_product_in_purchase = '" + price + "' AND count_product_in_purchase = '" + count + "'", ConnectionToDB.DB);
+                delProd.ExecuteNonQuery();
+
+                delProd = new SQLiteCommand("DELETE FROM warehouse WHERE id_product = '"+ id +"'", ConnectionToDB.DB);
+                delProd.ExecuteNonQuery();
+
+                delProd = new SQLiteCommand("DELETE FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
+                delProd.ExecuteNonQuery();
+            }
+            else
+            {
+                SQLiteCommand delProd = new SQLiteCommand("DELETE FROM product_list_in_purchase WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND price_product_in_purchase = '" + price + "' AND count_product_in_purchase = '" + count + "'", ConnectionToDB.DB);
+                delProd.ExecuteNonQuery();
+            }
+           
             this.Close();
 
             ConnectionToDB.closeDB();
